comments working, but not dynamic
