+++ views/recordId.pug
@@ -13,12 +13,14 @@ block content
             p= record.description
         div()
             h2 comment section
-            form(action=`${record.id}/comments/${comments.id}` method="post" class="comments-form")
+            form(action=`/comments` method="post" class="comments-form")
                 input(type='hidden' name='_csrf' value=csrfToken)
+                input(type='hidden' name='recordId' value=record.id)
+                input(type='hidden' name='userId' value=record.User.id)
                 div(class='text-area')
-                    label(for='comment') Comment here:
-                    textarea(name="comment" type='text' id="comment", cols="30", rows="10")
-                button(type='submit' id=`newComment-btn-${comments.id}`) submit
-        div()
+                    label(for='description') Comment here:
+                    textarea(name="description" type='text' id="description", cols="30", rows="10")
+                button(type='submit' id=`newComment-btn`) submit
+        div(id="comments-container")
             for comment in comments
                 p= comment.description
